pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Edgar_Meva"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "confluent-kafka (>=2.10.1,<3.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "mistralai (>=1.8.2,<2.0.0)",
    "pyspark (>=4.0.0,<5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "networkx (>=3.5,<4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pendulum (>=3.1.0,<4.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "pyldavis (>=3.4.1,<4.0.0)",
    "apache-airflow (>=3.0.2,<4.0.0)",
    "apache-airflow-providers-apache-spark (>=5.3.0,<6.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
